+++ needle/core/framework/cli.py
@@ -51,7 +51,7 @@ class CLI(Framework):
         self.register_option('port', Constants.GLOBAL_PORT, True, 'Port of the SSH agent on the testing device (needs to be != 22 to use USB)')
         self.register_option('username', Constants.GLOBAL_USERNAME, True, 'SSH Username of the testing device')
         self.register_option('password', Constants.PASSWORD_MASK, True, 'SSH Password of the testing device')
-        self.register_option(Constants.PASSWORD_CLEAR_NAME, Constants.GLOBAL_PASSWORD, True, 'SSH Password of the testing device')
+        self.register_option(Constants.PASSWORD_CLEAR, Constants.GLOBAL_PASSWORD, True, 'SSH Password of the testing device')
         self.register_option('pub_key_auth', Constants.GLOBAL_PUB_KEY_AUTH, True, 'Use public key auth to authenticate to the device. Key must be present in the ssh-agent if a passphrase is used')
         self.register_option('debug', Constants.GLOBAL_DEBUG, True, 'Enable debugging output')
         self.register_option('verbose', Constants.GLOBAL_VERBOSE, True, 'Enable verbose output')